fastapi/app/otlp.py
from opentelemetry import trace
from opentelemetry.exporter.otlp.proto.grpc.trace_exporter import OTLPSpanExporter
from opentelemetry.instrumentation.fastapi import FastAPIInstrumentor
from opentelemetry.instrumentation.logging import LoggingInstrumentor
from opentelemetry.sdk.resources import Resource
from opentelemetry.sdk.trace import TracerProvider
from opentelemetry.sdk.trace.export import BatchSpanProcessor

from starlette.types import ASGIApp


def setting_otlp(app: ASGIApp, endpoint: str, service_name: str, compose_service: str, log_correlation: bool = True) -> None:
    # set the service name to show in traces
    resource = Resource.create(attributes={
        "service.name": service_name,
        "compose_service": compose_service,
    })

    tracer_provider = TracerProvider(resource=resource)
    exporter = OTLPSpanExporter(endpoint=endpoint)
    processor = BatchSpanProcessor(exporter)
    tracer_provider.add_span_processor(processor)
    trace.set_tracer_provider(tracer_provider)

    if log_correlation:
        LoggingInstrumentor().instrument(set_logging_format=True)

    FastAPIInstrumentor.instrument_app(app, tracer_provider=tracer_provider)


def get_tracer():
    return trace.get_tracer(__name__)
