fastapi/requirements.txt
httpx==0.27.0
fastapi==0.111.0
uvicorn==0.29.0
opentelemetry-distro==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-instrumentation-logging==0.45b0
opentelemetry-exporter-otlp==1.24.0
starlette-prometheus
prometheus-client
# prometheus-client==0.20.0
tritonclient[all]==2.42.0
grpcio==1.60.0
grpcio-tools==1.59.3